Fix failing tests on Windows.
Test "test_vim.test_command" was failing because of the call to
os.unlink(..), as the file was still open in vim. Adding a conditional
to execute the call only if we are not running under Windows was the
most straight-forward "fix".

Test "test_vim.test_chdir" was failing for two reasons. Firstly because
of missing Windows-specific handling of path names in
find_file_in_path_option in file_search.c in Neovim core - this was
fixed in neovim/neovim#4711. Secondly because although one can indeed
`:cd /` on Windows (or at least, should be able to), a corresponding
call to `:pwd` will echo the root drive of the current working
directory of Neovim.