vim.buffers is a map not a list.
python-buffers describes vim.buffers as a mapping object. python-buffer
documents that b.number where b is a Buffer object can be used as a
python-buffers key. I.e. b == vim.buffers[b.number] should be true if b
is a valid buffer object.

nvim.py was implementing vim.buffers as a RemoteSequence which doesn't
provide the aforementioned properties. This change implements it as a
Buffer object instead. The latter is a new class that implements the
hybrid sequence/map interface that python-buffers describes.