change indexing-style in Range-Object-creation
This commit changes the indexing-style when creating a Range-object
from vim-style to Python-style, i.e. creating a Range-object spanning
the first three lines of a buffer changes from Range(buffer, 1, 3) to
Range(buffer, 0, 3), making the notation consistent with
Python-style-indexing and the behavior of the Buffer-object, that
already follows Python-style indexing.