Check count for 'not None' instead of truthy in command decorator
The command decorator was only checking the truthyness of `count`, which
meant that a value of 0 was considered falsy and so the value not passed
to the decorated function. A value of 0 for `count` should be valid
since this is the default in both vim and neovim (see `help v:count`).