Python-client 0.2.2
In this release support of python3.3 is dropped. Henceforth we want python3
rplugins to be able to assume the usage of asyncio, so they can use the asyncio
event loop and libraries that build on it.

Furthermore, a close() method is added on nvim session objects. When used as
a library for externally connecting to a nvim instance (i e not rplugins),
it is recommended to call the close() method on the session object when it is
not needed anymore. Alternatively, sessions can be used as a context manager:

    with neovim.attach('socket', path=thepath) as nvim:
        # do stuff with nvim session in this block:
        print(nvim.funcs.getpid())
        print(nvim.current.line)

This will close the session automatically.

Changes since 0.2.1:
* 2689ddc add tests for plugin decorators #298
* 63f257f allow library users to properly cleanup the event loop #303
* 59c184f expose the asyncio event loop as nvim.loop (python 3.4+ only) #294