Pynvim 0.3.1
This release renames the the package name (for PyPI and importing) to `pynvim`.
`import neovim` is still supported as an alias, but new API users should use
`import pynvim`.

If you already have installed `neovim` in pip, the best way to upgrade this time
is to upgrade the `neovim` package:

`pip install --upgrade neovim`

This will replace the neovim package with a dummy transition package, and
install pynvim as a dependency. For a new install, the prefered way is

`pip install pynvim`.

This release is otherwise functionally identical to pynvim 0.3.0, except for an
improved error message.

Changes since 0.3.0:
    - 1fcc17f More informative error message for usage from non-main thread
    - f237238 Rename package to `pynvim`