tests: remove cleanup_func (#414)
The "vim" fixture uses a new session always.

It was added when migrating to pytest in b65f62d, but is not really
necessary.  It could still be used to e.g. ensure tests are cleaning up
after themselves, but this is also not necessary really.

Adjusts existing tests, mainly with regard to handle IDs.