packaging: add pynvim.__version__ attribute (0.5.0.dev0)
pynvim previously did not have the `pynvim.__version__` attribute,
which is a popular convention for python modules.

We add the `pynvim.__version__` string in addition to `pynvim.VERSION`,
and let `pynvim._version` be the only file that contains the version
metadata for the pynvim package. `setup.py` can also get the version
string from this file.

pynvim version is bumped up to a devel version: 0.5.0.dev0 (from 0.4.3).