refactor: remove the use of forward references
With PEP-563 (Postponed evaluation of annotations), we can have a more
clean syntax of writing type hints, not using 'forward references' to
names that have not been defined yet in the form of string literals.

Note that Python 3.7+ (the minimum python version that is currently
supported by pynvim) supports PEP-563.