fix(test_options): adapt global-local option to new Nvim default
The test previously depended on the 'statusline' option being empty by
default.  The default 'statusline' value changed in Nvim commit
28e31f5d3d16fac349d5e2b55837afddb822b0f3 (2025-04-22):

    feat(options): default statusline expression #33036

Adjust the test to first read the default (global) value of
'statusline', then change the local value of the option, then re-read
the global value and verify it hasn't changed.